* Source/cairo/CairoGState.m           abs -> fabs         * Source/x11/XGGLContext.m           format string fix         * Headers/x11/XGOpenGL.h           glxminorversion is int, not long int
Looks good, Ivan Vucica



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@38554 72102866-910b-0410-8b05-ffd578937521

+++ Headers/x11/XGOpenGL.h
@@ -57,7 +57,7 @@
 @interface XGGLPixelFormat : NSOpenGLPixelFormat
 {
   Display * display;
-  long int glxminorversion;
+  int glxminorversion;
 
   GLXFBConfig  *fbconfig;
   int pickedFBConfig;
